# openapi/v1/api.yaml
openapi: 3.0.0
paths:
...
  /v1/vehicles:
    get:
      operationId: VehiclesController_findAll
      summary: List all vehicles with cursor-based pagination
      parameters: []
      responses:
        '200':
          description: ''
      tags:
        - vehicles
      security:
        - bearer: []
    post:
      operationId: VehiclesController_create
      summary: Create a new vehicle
      parameters: []
      responses:
        '201':
          description: ''
      tags:
        - vehicles
      security:
        - bearer: []
  /v1/vehicles/{id}:
    get:
      operationId: VehiclesController_findById
      summary: Get vehicle by ID
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        '200':
          description: ''
      tags:
        - vehicles
      security:
        - bearer: []
    patch:
      operationId: VehiclesController_update
      summary: Update vehicle
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        '200':
          description: ''
      tags:
        - vehicles
      security:
        - bearer: []
    delete:
      operationId: VehiclesController_delete
      summary: Soft delete a vehicle
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        '204':
          description: ''
      tags:
        - vehicles
      security:
        - bearer: []
  /v1/vehicles/{id}/status:
    patch:
      operationId: VehiclesController_updateStatus
      summary: Update vehicle status
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        '200':
          description: ''
      tags:
        - vehicles
      security:
        - bearer: []
  /v1/vehicles/{id}/availability:
    get:
      operationId: VehiclesController_checkAvailability
      summary: Check vehicle availability
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        '200':
          description: ''
      tags:
        - vehicles
      security:
        - bearer: []
...
info:
  title: Gestão Fretamento Pro API
  description: API REST para gestão de fretamento empresarial
  version: 1.0.0
  contact: {}
...
servers: []
components:
  securitySchemes:
    bearer:
      scheme: bearer
      bearerFormat: JWT
      type: http
...